# openapi.yaml
openapi: 3.0.3
info:
  title: PadelZGZ API
  description: API REST para gestión de clubes de pádel, reservas de pistas, torneos e inscripciones en Zaragoza.
  version: 1.0.0
  contact:
    name: PadelZGZ
    email: info@padelzgz.com
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReservaInDTO:
      type: object
      required: [fecha, horaInicio, horaFin, pistaId, usuarioId]
      properties:
        fecha:
          type: string
          format: date
          example: "2025-07-20"
        horaInicio:
          type: string
          example: "10:00"
        horaFin:
          type: string
          example: "11:00"
        precio:
          type: number
          format: float
          example: 12.50
        pagado:
          type: boolean
          example: false
        comentario:
          type: string
          example: "Reserva para partido amistoso"
        pistaId:
          type: integer
          format: int64
          example: 1
        usuarioId:
          type: integer
          format: int64
          example: 1
...
paths:
...
  /reservas:
    get:
      tags: [Reservas]
      summary: Listar reservas con filtros
      security:
        - bearerAuth: []
      parameters:
        - name: fecha
          in: query
          schema:
            type: string
            format: date
        - name: pagado
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de reservas
        '500':
          description: Error interno

    post:
      tags: [Reservas]
      summary: Crear reserva (usa DTO con IDs de pista y usuario)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaInDTO'
      responses:
        '201':
          description: Reserva creada
        '400':
          description: Datos inválidos
        '404':
          description: Pista o usuario no encontrado
        '500':
          description: Error interno
...